+++ lib/src/jar/persist.dart
@@ -10,7 +10,7 @@ import 'default.dart';
 /// [PersistCookieJar] is a cookie manager which implements
 /// the standard cookie policy declared in RFC.
 /// [PersistCookieJar]  persists the cookies in files, if the application exit,
-/// the cookies always exist unless user explicitly called [delete].
+/// the cookies always exist unless user explicitly deleted with [deleteWhere].
 class PersistCookieJar extends DefaultCookieJar {
   /// [persistSession] is whether persisting the cookies that without
   /// "expires" or "max-age" attribute.
@@ -157,6 +157,7 @@ class PersistCookieJar extends DefaultCookieJar {
   /// This API will delete all cookies for the `uri.host`, it will ignored the `uri.path`.
   ///
   /// [withDomainSharedCookie] `true` will delete the domain-shared cookies.
+  @Deprecated('Use deleteWhere instead')
   @override
   Future<void> delete(Uri uri, [bool withDomainSharedCookie = false]) async {
     await _checkInitialized();
@@ -169,6 +170,15 @@ class PersistCookieJar extends DefaultCookieJar {
     if (withDomainSharedCookie) {
       await storage.write(_domainsKey, json.encode(domainCookies));
     }
+  }
+
+  @override
+  Future<void> deleteWhere(bool Function(Cookie cookie) test) async {
+    await _checkInitialized();
+    super.deleteWhere(test);
+
+    await storage.write(_indexKey, json.encode(_hostSet.toList()));
+    await storage.write(_domainsKey, json.encode(domainCookies));
   }
 
   /// Delete all cookies files in the [storage] and the memory.